Fixing SonarCLoud issues, next round

+++ src/raspberrypi/rasctl_commands.h
@@ -48,7 +48,7 @@ private:
 
 	PbCommand command;
 	string hostname;
-	uint16_t port;
+	int port;
 	string token;
 	string locale;
